feat: event persistence consumer utility scripts
